Improved consistency for receive_packet
Right now packets that are received with the default protocol are associated with a link that belongs to the sending node. But packets that are received over a custom protocol are associated with a link that belongs to the receiving node. This change makes the link always belong to the sending node, regardless of protocol.